backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pymongo==4.10.1
motor==3.6.0
httpx==0.27.2
email-normalize==2.0.0
beautifulsoup4==4.12.3
tldextract==5.1.2
spacy==3.7.6
scikit-learn==1.5.2
faiss-cpu==1.8.0.post1
numpy==1.26.4
ollama==0.3.3
joblib==1.4.2
feedparser==6.0.11
cachetools==5.3.3
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
resend==2.4.0
certifi==2024.2.2
